Add files via upload
Assignment: CRISP Fellow  
 
Guidelines 
•	Use the nomenclature: FirstName_LastName_CRISP • 	Kindly be brief and professional. 
•	Please attempt BOTH the questions in the assignment given below. 
•	The quality of thought process matters more than the content of the recommendation/output 
•	You are free to make assumptions for information that’s not given here, but please explicitly state them in your analysis. 
•	Cite your sources/ references. No plagiarism please. 
Please upload your responses to BOTH the questions (presentation/pdf formats only) in a single file on the Online application portal. 
 
QUESTION 1: CASE STUDY 
Out of the total 1200 tribal habitations in a State, schools were there only in 824 habitations. Even in the existing schools, teacher truancy has been a major issue, resulting in the high dropout rate of more than 80 from class 1 to 5.  
In order to address the problem, the Govt has decided to employ educated tribal youth in the villages as teachers. As a result, schools were set up in all the 1200 habitations, with tribal youth, who were educated up to 10th class even though they were not trained in teaching. All of them have been given the normal salary of the regular teachers.  
Three years later, when the system was checked, it was found that the drop-out rate had not improved, and the teacher truancy remained the same. All the tribal youth appointed as teachers shifted their residence to the nearest town.  
You are required to create a presentation (not exceeding 10 slides), critically examining the policy of Govt in engaging untrained tribal youth as teachers. Do you agree with the Policy? If so, why? If not, what could have been done differently?  
 
QUESTION 2: ANALYTICAL QUESTION 
Note: Candidates should collect relevant data from secondary sources and base the answer on the analysis of the data along with suitable data visualization.  
“India’s Economic Growth has a far greater impact on poverty alleviation than measures to reduce inequality”. Examine critically. Do you agree with the statement? If so, why? If not, give reasons and your suggestions. 
 In solution to this question, you are required to create a short presentation (not exceeding 5 slides).